Add files via upload
In this we a take a very large dataset in which there are some approx 67 lac rows, in this we have to predict the tip given by the customer. In this project we also compare the Snap ML with the scikit Learn library, and we conclude that on the very large data set the accuracy of the both library is approximately same but the time taken to process data, fit data is less in the snap ML as compare to the scikit learn. 
Snap ML is a high-performance IBM library for ML modeling. It provides highly-efficient CPU/GPU implementations of linear models and tree-based models. Snap ML not only accelerates ML algorithms through system awareness, but also offers novel ML algorithms with best-in-class accuracy. For more informatio